sched: Disable gentle fair sleepers
Sleeper Fairness is a concept used by CFS which treat sleeping/waiting
tasks as if they were in a run queue. This implies tasks which spend
most of the time waiting for an user input and such will get a fair
share of CPU when they need it. Disabling Gentle Fair Sleepers could
improve UI responsiveness.

Change-Id: I35c3184ddeb927aab1dba01e2c2eedb32769434d
Signed-off-by: Albert I <krascgq@outlook.co.id>