PM / devfreq: Add a governor white list
Many of the devfreq governors compiled in are unstable/incomplete, so we
don't want users switching to them. Compiling out the unused governors is
messy because many other drivers link to them, and there are not proper
 #ifdefs in place.

Add a white list of the only governors intended for use to the governor
sysfs node in order to prevent users from enabling governors that unstable.

Change-Id: I50ade60acc11b99efeee964d17c33a10d3ad7c4b
Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>