base: dd: Limit probe deferral when booted into recovery
Commit 11f58b3ece770c32f90ad510713644dcee3f40f2 "base: dd: Continue probe deferral forever as needed
after late_initcall" may be able to fix some rare issues with the audio
card not probing in time and causing the boot process to hang on waiting
for the audio card to become available when booting Android, but it
causes constant background errors in TWRP where the system does not
properly initiate probing of the card. As a workaround, disable this
functionality when booted into recovery as recovery will not be using
sound anyway, so it doesn't matter if it's broken.

The aforementioned errors from dmesg:
[   59.947272] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.037756] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.128177] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.218596] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.309023] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.399271] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.487948] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.576135] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered
[   60.664564] sdm845-asoc-snd soc:sound-tavil: ASoC: platform (null) not registered

Signed-off-by: kdrag0n <dragon@khronodragon.com>